requirements.txt
fastai>=2.7.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
jupyter>=1.0.0
ipywidgets>=7.6.0

# Web interface
streamlit>=1.28.0
flask>=2.3.0
werkzeug>=2.3.0

# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Additional utilities
tqdm>=4.65.0
opencv-python>=4.8.0
